+++ exporters/otlp/src/otlp_metric_utils.cc
@@ -229,6 +229,7 @@ void OtlpMetricUtils::PopulateResourceMetrics(
     proto::common::v1::InstrumentationScope *scope = scope_lib_metrics->mutable_scope();
     scope->set_name(scope_metrics.scope_->GetName());
     scope->set_version(scope_metrics.scope_->GetVersion());
+    resource_metrics->set_schema_url(scope_metrics.scope_->GetSchemaURL());
 
     for (auto &metric_data : scope_metrics.metric_data_)
     {