+++ sdk/test/metrics/aggregation_test.cc
@@ -20,8 +20,8 @@ TEST(Aggregation, LongSumAggregation)
   auto sum_data = nostd::get<SumPointData>(data);
   ASSERT_TRUE(nostd::holds_alternative<int64_t>(sum_data.value_));
   EXPECT_EQ(nostd::get<int64_t>(sum_data.value_), 0);
-  aggr.Aggregate((int64_t)12, {});
-  aggr.Aggregate((int64_t)0, {});
+  aggr.Aggregate(static_cast<int64_t>(12), {});
+  aggr.Aggregate(static_cast<int64_t>(0), {});
   sum_data = nostd::get<SumPointData>(aggr.ToPoint());
   EXPECT_EQ(nostd::get<int64_t>(sum_data.value_), 12);
 }
@@ -48,8 +48,8 @@ TEST(Aggregation, LongLastValueAggregation)
   auto lastvalue_data = nostd::get<LastValuePointData>(data);
   ASSERT_TRUE(nostd::holds_alternative<int64_t>(lastvalue_data.value_));
   EXPECT_EQ(lastvalue_data.is_lastvalue_valid_, false);
-  aggr.Aggregate((int64_t)12, {});
-  aggr.Aggregate((int64_t)1, {});
+  aggr.Aggregate(static_cast<int64_t>(12), {});
+  aggr.Aggregate(static_cast<int64_t>(1), {});
   lastvalue_data = nostd::get<LastValuePointData>(aggr.ToPoint());
   EXPECT_EQ(nostd::get<int64_t>(lastvalue_data.value_), 1.0);
 }
@@ -77,8 +77,8 @@ TEST(Aggregation, LongHistogramAggregation)
   ASSERT_TRUE(nostd::holds_alternative<int64_t>(histogram_data.sum_));
   EXPECT_EQ(nostd::get<int64_t>(histogram_data.sum_), 0);
   EXPECT_EQ(histogram_data.count_, 0);
-  aggr.Aggregate((int64_t)12, {});   // lies in third bucket
-  aggr.Aggregate((int64_t)100, {});  // lies in sixth bucket
+  aggr.Aggregate(static_cast<int64_t>(12), {});   // lies in third bucket
+  aggr.Aggregate(static_cast<int64_t>(100), {});  // lies in sixth bucket
   histogram_data = nostd::get<HistogramPointData>(aggr.ToPoint());
   EXPECT_EQ(nostd::get<int64_t>(histogram_data.min_), 12);
   EXPECT_EQ(nostd::get<int64_t>(histogram_data.max_), 100);
@@ -86,8 +86,8 @@ TEST(Aggregation, LongHistogramAggregation)
   EXPECT_EQ(histogram_data.count_, 2);
   EXPECT_EQ(histogram_data.counts_[3], 1);
   EXPECT_EQ(histogram_data.counts_[6], 1);
-  aggr.Aggregate((int64_t)13, {});   // lies in third bucket
-  aggr.Aggregate((int64_t)252, {});  // lies in eight bucket
+  aggr.Aggregate(static_cast<int64_t>(13), {});   // lies in third bucket
+  aggr.Aggregate(static_cast<int64_t>(252), {});  // lies in eight bucket
   histogram_data = nostd::get<HistogramPointData>(aggr.ToPoint());
   EXPECT_EQ(histogram_data.count_, 4);
   EXPECT_EQ(histogram_data.counts_[3], 2);
@@ -97,16 +97,16 @@ TEST(Aggregation, LongHistogramAggregation)
 
   // Merge
   LongHistogramAggregation aggr1;
-  aggr1.Aggregate((int64_t)1, {});
-  aggr1.Aggregate((int64_t)11, {});
-  aggr1.Aggregate((int64_t)26, {});
+  aggr1.Aggregate(static_cast<int64_t>(1), {});
+  aggr1.Aggregate(static_cast<int64_t>(11), {});
+  aggr1.Aggregate(static_cast<int64_t>(26), {});
 
   LongHistogramAggregation aggr2;
-  aggr2.Aggregate((int64_t)2, {});
-  aggr2.Aggregate((int64_t)3, {});
-  aggr2.Aggregate((int64_t)13, {});
-  aggr2.Aggregate((int64_t)28, {});
-  aggr2.Aggregate((int64_t)105, {});
+  aggr2.Aggregate(static_cast<int64_t>(2), {});
+  aggr2.Aggregate(static_cast<int64_t>(3), {});
+  aggr2.Aggregate(static_cast<int64_t>(13), {});
+  aggr2.Aggregate(static_cast<int64_t>(28), {});
+  aggr2.Aggregate(static_cast<int64_t>(105), {});
 
   auto aggr3     = aggr1.Merge(aggr2);
   histogram_data = nostd::get<HistogramPointData>(aggr3->ToPoint());